+++ anchor/programs/solstarter/src/state/project.rs
@@ -15,8 +15,8 @@ pub struct Project {
     #[max_len(3000)]
     pub project_description: String,
 
-    pub goal_amount: u32,
-    pub raised_amount: u32,
+    pub goal_amount: u64,
+    pub raised_amount: u64,
     pub created_time: i64,
     pub end_time: i64,
     pub status: Status,
